Unsuccessful attempt to return sent message as XML stream to application.
We will implement dowload first and then we will decide whether to invent some
common format (similar do XSD type tReturnedMessage or to use ESS format), or
whether we pass it to application. Nevertheless we are establishing struct
isds_message which is intended to carry all data reletated to any message.