Improve time zone offset computation.
Unfortunatelly, this does not consider DST. Global daylight is obsolete and it
does not show current state. There is gettimetofday()' tz_dsttime.

But it still introduce race, when DST changes between preparing
from_time/to_time arguments and when they are converted to string.

The best way will be to move to UTC struct timeval. This also allows to
express fraction of second which is allowed by ISO and which ISDS uses.