GetMessageAuthor implemented
This service is exported as get_message_sender() public function. It used new
type isds_sender_type that is probably superset of isds_UserType, however not
in semantic way (the SENDERTYPE_VIRTUAL value is not a user actually).  Also
because XML schema allows any string, I decided to pass raw value to
application too to application to cope with unkown values.