Base64 codec uses proper types
Types size_t and int8_t are used instead of int and char. The char has been
abandoned because is can occupy more than one byte and I think Base64 is
specified as sequence of octets not as a of characters.

Little drawback is using int8_t is supported on system with 8-bit byte only.

In addition, small off-by-one error in decodinging invalid input has been
fixed.

Coding style has been adjusted to the rest of this library.