Remove a use-after-free when parsing SOAP response
_isds_soap() returned a copy of an XML node list with content of the SOAP
Body element. However, the copy still linked to the original XML document
dictionary and the original XML document was deallocated before leaving
_isds_soap().

This patch changes the API of the internal _isds_soap() function to return
original XML document and a pointer to the content of the SOAP Body.

So the problem moves to the upper layer. Namely _isds() and _czp_czpdeposit().

I'm not sure the the two upper functions build a new XML document from the
SOAP Body content properly, maybe it introduces some memory leaks. It will
need reading libxml2 sources more thoroughly.