Silent a signed-comparison warning in Base-64 decoding
The sizeof() is used to compute a number of members of a static array.
The unsigned (size_t) value was then used to compare against a signed
value. And here compiler can warn.

To prevent from pointless sign extension to size_t width, I decided to
store the "constant" (it's 80) into int8_t as it fits there.