Regenerate certificates with longer expiration time
Certificates in server/tls directory are used when running simline tests.
The certificates expired at the end of August 2015, therefore
certificate_user_password_authentication test unit failed.

This fixes the tests by providing new certificates that are valid for next
50 years.

In the future, we will probably generate the certificates on the fly. This is
only a hotfix.

The new certificates have reversed order of fields in the distinguished names
(because GnuTLS' certtool does so and I'm lazy to generate them again).
Therefore the distinguished name of the client was adjusted in the test.