Format tv_usec as int32_t in timeval2timestring()
Windows' PRIdMAX is not compatible with MinGW's printf(). But because we know
the tv_usec is not bigger than 999999, we can handle it as int32_t.

OTOH this is very bad because libisds is written in ISOC99 and toolchains not
supporting 16 years old standard deserve eternal damnation. Therefore I will
not adjust tests only to conform broken compilers. Actually maybe I will
revert this change completly in the future beacuse I'm not fully convinced.