Fixed a potential bug when printing debugging information.
According to C standard, the string precision/length limit argument of printf()
when using "%.*s" must have int type. Using size_t instead can cause wrong
interpretation of data stored on stack. Depending on the compiler and system
this may result in random programme crashes. It is better to stick to
the standard and pass int as string length.