Do not use dolar in formatting strings
ISO C 99 does not support "$" to control print(3) arguments order. This is not
support on mingw platform.

This patch changes the library code in order not to use it. It sucks because
it makes the code less elegant and less readable.

The same problem can occur in translated messages. I worry translations
without capability of words reordering will be dumb. I urge people to use
better toolchains.