test: Allow skipping tests
This patch extends testing library to allow skipping tests.

A new SKIP_TEST(reason) macro can be used to exit a test function because
a current test cannot proceed for outer reasons.

A new SKIP_TESTS(number, reason) macro can be used to skip a number
of tests. The skipped tests must not be invoked by a program control
flow. SUM_TEST() still must be called at the end of the unit. This is
different from ABORT_UNIT() that reports a unit failure.