Test: Accept IE_NETWORK error when client does not provide a correct certificate
After upgrading GnuTLS to 3.6.4,
test/simline/certificate_user_password_authentication tests that exhibit
a missing client's certificate and a bad client's certificate started to
fail. That's because cURL returns a network error instead of a TLS error in
that case when TLSv1.3 is used.

This patch work arounds this cURL deficiency.

See a similar GnuTLS bug report
<https://gitlab.com/gnutls/gnutls/issues/615>.